refactor: share test client audio helpers and document websocket transport #23
